should work

+++ api/queue.js
@@ -105,13 +105,17 @@ export default async (req, res) => {
         "SELECT * FROM queue"
     )
     
-
-    const selectResponse3 = await pool.query(
-        "DELETE FROM matches WHERE (playerone = $1 OR playertwo = $1) AND winner IS NULL",
+    const selectResponse5 = await pool.query(
+        "SELECT * FROM queue WHERE username = $1",
         [userName]
-    );    
+    )
 
-    if (selectResponse2.rows.length > 0) {
+    if (selectResponse2.rows.length > 0 && selectResponse5 === 0) {
+
+        const selectResponse3 = await pool.query(
+            "DELETE FROM matches WHERE (playerone = $1 OR playertwo = $1) AND winner IS NULL",
+            [userName]
+        );
 
         const selectResponse4 = await pool.query(
             "DELETE FROM queue WHERE username = $1",
